Add tests for filter design
This confirms the correctness of our results and compares accuracy vs.
MATLAB for all but the bilinear step in digital filter design. I am a
little suspicious that bilinear step is not particularly numerically
precise, but I have no good way of testing that at the moment, since I
can't perform the calculations with extended precision.