Faster FIR filtfilt
For FIR filters, it is better to use firfilt than filt since it ensures
an appropriately fast algorithm is used. Additionally, the filter can
simply be convolved with itself so that it is not necessary to process
the data in the forwards and backwards direction. There is no need to
compute the initial state because it has no effect on the result after
the signal is padded.

This algorithm could still be made more efficient with an in-place
variant of firfilt that can save the FFT of the filter.