Fix inference of return type of spectrogram
The instability of the output type doesn't make a huge performance
difference unless there are a large number of segments, but caused the
return type of spectrogram to be inferred as
Union(Spectrogram{Float64,Frequencies},Spectrogram{Complex{Float64},Frequencies}))
which could be bad for user code that wants to process it further.

I removed the psdonly argument from the docs since I don't think
there's much reason to use it directly.