Create mutating versions of fftfilt and filt
I have split `fftfilt` into two functions, the original `fftfilt` which
initializes the output array, `ffiltfilt!` which checks the size of the output
array is consistent with the input array, and `_fftfilt!` which performs the
filtering by mutating the output array without checking its inputs.

Similarly, I have separated the FIR method of `filt` into the original function
which initializes the output array, `filt_choose_alg!` which checks if the
output array and input array are the same size, and `_filt_choose_alg!` which
will choose to use fftfilt or time-domain filtering based on the same heuristics.