Revert "Align behavior of filt and filt! for FIR filters"
This reverts commit 0e8b6afbf2a58a5115e0d26ab5039b8f567f9217.

`fftfilt!` does not work if the output array is an alias of itself. As `fftfilt`
uses the overlap add algorithm, fixing this may be difficult. In contrast,
`filt!` can mutate its own input when the output is an alias of the input
according to the documentation. However, this only works with the time-domain
algorithm. Therefore, changing the behavior of `filt!` to choose between fft
filtering and time-domain filtering makes its behavior inconsistent with the
documentation.