Add @simd macro to amenable loops
A number of loops that satisfy the constraints to use the @simd macro did not. I
have added them to the loops that I could find, and found that performance and
memory use for `mt_pgram` increased by a few percent. While I expect other
performance gains to be similarly modest, it's a 'free' boost.