Generalize optimalfftfiltlength to arbitrary input sizes
As described in #213, FIR filtering will fail if the input vector is larger than
2^28, which is the last saved benchmark used by optimalfftfiltlength to
heuristically choose a chunk size L for the overlap-add algorithm used in fft
filt.

Instead of relying on saved benchmarks to pick L, it is possible to calculate
the per-sample complexity of the overlap-add algorithm, as a function of nfft
and by extension L [0]. I have replaced looking up benchmarks with calls to a
new function, oa_fft_complexity, which calculates the per-sample complexity for
a given nfft that is a power of 2 and the filter size nb. I then choose the
minimum complexity for a range of nffts, and use that as the best nfft.

[0] e.g. https://www.allaboutcircuits.com/technical-articles/overlap-add-method-linear-filtering-based-on-the-discrete-fourier-transform/