Update index.rst with new documentation link
DSP.jl's legacy documentation hosted on readthedocs points to the current documentation, which is hosted by github. Changing the travis build script in 2675206e85edd7409c20c0187f6af867f9d9e29c caused the URL of the current documentation to change, but this change was not added to the legacy documentation hosted on readthedocs.