Avoid InexactErrors when constructing PolynomialRatio with integers
This fixes #142 by making an inner constructor for integers that checks that
coefficients are normalized, but does not attempt to normalize them. An outer
constructor handles automatic normalization of integer coefficients, which
converts them to floating point coefficients. To construct PolynomialRatio with
integer coefficients, already normalized coefficients, use
PolynomialRatio{<:Integer(...).