Make xcorr padding optional
As brought up in #253, xcorr currently has the unintuitive behavior of doing
padding in addition to the padding done by `conv`. This means that for inputs of
different lengths, `xcorr` results are not the same length as `conv`, and that
the beginning of the result is filled with zeros. While Numpy and Matlab both
return padded results, they both fill the end of result with zeros, instead of
the beginning.

This commit changes the default behavior to be more in line with the
mathematical definition of cross-correlations, and also makes `xcorr` more in
line with the behavior of `conv`. This new behvaior can be disabled by
setting the new `mode` keyword to `:legacy`. Deprecation warnings are thrown if
the result of `xcorr` would be altered by the new behavior.