Merge code paths for 1D and ND convolutions
Zero-padding is the one part of the convolution process that benefits from being
handled differently with 1D inputs and ND inputs. I have extended the _zeropad
function to do zero-padding in arbitrary dimensions in a type stable way. With
this method in place, then the rest of the convolution algorithm is the nearly
the same between vectors and ND arrays.