Make xcorr padding optional (#266)
* Make xcorr padding optional

As brought up in #253, xcorr currently has the unintuitive behaviour of doing
padding in addition to the padding done by `conv`. This means that for inputs of
different lengths, `xcorr` results are not the same length as `conv`.

This commit allows users to select the pad behavior, with `:none` being
in line with mathematical definition of cross-correlations, as well as `conv`. 
The old behaviour can be selected with `:longest`. Deprecation warnings are thrown
if no keyword is specified, in preparation for changing the default behaviour in a
future commit.