Release unused memory when using conv
`resize!` changes the length of an array, however it does not free the
underlying unused memory if the array is reduced in length. `sizehint!` will
free the memory corresponding to the portion of the array dropped by `resize!`.

`conv` drops the last portion of the result, corresponding to the padded portion
of the result. By using `sizehint!` after resizing the result, the padding is
freed from memory.