Make conv work for more element types
Previously, conv would fail for inputs with differing floating point elements
types. This commit fixes this problem by adding an extra promotion method.

In addition, this commit allows other numbers that are not BLAS floats to be
used. Due to the buffering invovled, and reuse of buffers for the different
inputs, it is still necessary to have the core `conv` kernal operate on BLAS
floats, but more types of inputs can now be accepted and converted.