Allow cross correlations to work with arbitrary offsets and padding
Cross-correlation of input arrays with arbitrary indexing did not work in
conjunction with padding, because the padding did not preserve the indexing.
Adding a zero-padding function that preserves indexing, `_zeropad_keep_offset`,
fixes this problem. The `_zeropad` function works the same as before, and
returns an array with base-1 indexing. I have added `_zeropad!` methods that
also work with arbitrary indexing for the padded array, which are dispatched
based on the axes of the of padded array.

Fixes #288