Simplify correction of using a larger FFT than necessary in conv
This commit simplifies the correction of using a larger fft than necessary while
doing convolution, as per @HDictus' suggestion. This can occur in
`optimalfftfiltlength`, which searches `nfft` over powers of two to minimized
the total complexity of the convolution. However, when `nfft` is larger than the
full convolution, it makes sense to consider other fast powers. This condition
was previously detected based on the output of a fft convolution being larger
than the input. However, in this same condition, `nfft` is also larger than the
full convolution size, `nfull`. Because `nfft` and `nfull` are already computed,
it is faster to not calculate the output size to detect this condition.