Add `using DSP` to all doctests
Some doctests were failing because `using DSP` was not specifically included in
the example. I have added a setup step in doc/make.jl that includes the DSP
module in all doctests, so that examples that do not explicitly include `using
DSP` will still be able to use DSP functions.