Relax type bounds of conv conversion method
`Conv` currently only works with floating point numbers, and converts arrays to
have the same dimensions and floating point element type prior to convolution.
There are a number of conversion methods for `conv` that perform this
conversion. However, if the two input arrays do not have the same element type,
and further if one of the inputs has a non-floating point real element type, for
example a fixed point number, then there is currently no conversion method that
will match this case. This results in a stack overflow. To fix this, I have
relaxed the type bounds of the method `conv(::AbstractArray{T, N},
::AbstractArray{T, N}) where {T<:Number, N}` so that it no longer requires the
two input arrays have the same element type.