Better handling of gain in filter accuracy tests
When testing the accuracy of a filter in zero/pole/gain (zpk) form,
directly comparing the gain k may be misleading: As k depends on the
poles and zeros, deviation in those may have arbitrary effects on k. So
a filter that is less accurate in terms of poles and zeros may just
accidentally have better accuracy in k.

Therfore, this changes the `zpkfilter_accuracy` function to take an
optional keyword parameter `compare_gain_at`, which, if given, lets it
compare the filter gain computed at the given frequency (e.g. 0 for a
lowpass) instead of plain k. By using this in a few tests, it is
possible to tighten the error bounds.