Make `SecondOrderSections(::ZeroPoleGain)`more robust (#435)
* Make `SecondOrderSections(::ZeroPoleGain)`more robust

The helper function `groupzp` assumes that both zeros and poles are
ordered such that complex entries appear in conjugate pairs
consecutively and further, that these pairs are orderer consistently:
Either always the positive imaginary part first, or always the negative
imaginary part first. The output of `split_real_complex` fulfills this.
However, for the poles, there is an intermediate sorting step which
apparently can interfere with this ordering under certain circumstances.
This can lead to a `BoundsError` (#432), but also to zeros which are not
conjugate pairs being grouped together, thereby silently producing wrong
results. While #433 was a band-aid preventing the `BoundsError`, this
should fix the underlying issue by integrating the sorting into
`split_real_complex` in such a way that the required ordering is
ensured.

* Bump version to 0.7.2