Rework definition of `FloatRange` (fix for Julia 1.8) (#456)
In Julia 1.7, the definition of `StepRangeLen` acquired another type
parameter, making the fields typed with `FloatRange{Float64}` abstractly
typed. While that is unfortunate but not critical,
https://github.com/JuliaLang/julia/pull/43059 further broke `convert`ing
to this abstract type, causing failures within DSP.jl. This change
reworks the definition of `FloatRange` to pick the type of an appropriate
`range` call, making it adjust automatically to the changes in Julia. To
simplify things, the type parameter is dropped, as it was only used with
`Float64` anyway.