Shift frequency normalization from construction of FilterType to filter design
When designing a digital filter using function digitalfilter() the
sampling frequency is not passed to the FilterType definition anymore,
but to function digitalfilter(). That means frequency normalization is
not done during construction of FilterType anymore, but during filter
design inside function digitalfilter().

For this the sampling frequency has to be passed to the internal
functions prewarp(), firprototype() and scalefactor() as well.

The intended side effect is, that an analog filter can now be designed
by specifying analog frequencies directly without any frequency
normalization and without a sampling frequency.