Rework "TF to ZPK" test to cope with OpenBLAS update (#483)
OpenBLAS v0.3.21 gives slightly different eigen values, leading to
different polynomial roots, which made the old test fail. In particular,
the DC gain of the filter is now `0.9999999999999846`, while the Matlab
based coefficients yield `0.9999999999999951`. Thus the error is more
than twice that of the Matlab version, making the test fail. But of
course, both results are sufficiently close to 1 to be acceptable.

This replaces the test with an error across a larger frequency range
where indeed, the Matlab based version is still worse.