Simplify return value by omitting `converged`.
The `quinn` functions now only return a tuple with two values;
the first is the estimated frequency, and the other is a boolean
which is true if the algorithm ran for the maximum allowed number
of iterations. The `converged` return value is somewhat redundant
(and unavailable in the complex case).