Improve `jacobsen` frequency estimator for real inputs
Before this commit, the `jacobsen` function did not perform any
interpolation when the DFT peak occurred at frequencies 0 or
Nyquist (`fs/2`).

Using the symmetry of the DFT, the estimation formula can be both
improved (by interpolating in all cases, with no exceptions), and
simplified (since the DFT values around the peak are (often) complex
conjugates).

This is slightly tricky because, in order to save computation
effort, we use the `rfft` instead of the `fft`. The formula can
be simplified when the peak occurs at 0, and when it occurs at
Nyquist and `N` is even.

The result is overall improved error performance near zero and
near Nyquist, even though in some frequency ranges the error
increases.