Adding documentation for bilinear function
This commit adds documentation for the bilinear function in the
`Filters` module. docstrings are added in source code for the two
methods of bilinear function with the top level method explained
in detail to providde a clear explanation of what the transform
function is doing.

The function is also added to the export list in `Filters.jl` so that
it is available for Documenter and also for users to use.

The documentation is added to the `docs/src/filters.md` file in the
`Filters` module, Filter design section where the bilinear function
is added as a general digital filter design method.

No code has been changed in this commit, only documentation has been
added and an extra function has been added to the export list.