Enable Analog Filter Design (#458)
* Shift frequency normalization from construction of FilterType to filter design

When designing a digital filter using function digitalfilter() the
sampling frequency is not passed to the FilterType definition anymore,
but to function digitalfilter(). That means frequency normalization is
not done during construction of FilterType anymore, but during filter
design inside function digitalfilter().

For this the sampling frequency has to be passed to the internal
functions prewarp(), firprototype() and scalefactor() as well.

The intended side effect is, that an analog filter can now be designed
by specifying analog frequencies directly without any frequency
normalization and without a sampling frequency.

* Update documentation

* Extend filter design tests

Extend test case 'digital IIR' to test digital filter design with
non-normalized frequency values in the filter response definition and
specified sampling frequency.

* Fix test set "freq. scaling"

This is an amendment to commit 54bc9d55cd207233cfdd6764ce269c659c7e5ecd
where test sets have been introduced. The code following the current
test set "freq. scaling" is actually a "lowpass" test. It seems that a
corresponding comment on this low pass filter test was already missing
before this commit. And the comment "Freqency scaling" seems to be meant
for all four following test sets "lowpass", "highpass", "bandpass" and
"bandstop".

* Rename test set "freq. scaling" to "analogfilter"

The new test set name relates to the name of the tested function
`analogfilter()` and not to what it is doing.

* Remove unnecessary conversion of filter representation

In the current version the function `analogfilter()` already generates
a `ZeroPoleGain` filter representation. It seems, that in earlier
versions this wasn't the case (see e.g. commit
7fd21020f814cba54da87d39d9f71a3635f99394)

* Update analog filter design tests using non-normalized frequency values

For providing new test data I had no MATLAB available and therefore
switched to Octave. Unfortunatelly Octave's elliptic filter design
function `ellip()` generates different filter coefficients than Julia
when designing a 20th order analog filter. Maybe this is a bug in
Octave's 'signal' package. I changed the test to design a Chebyshev
type II filter instead.

---------

Co-authored-by: Christian Gruber <Christian.Gruber-e6t@rub.de>
Co-authored-by: Viral B. Shah <ViralBShah@users.noreply.github.com>