Use Bessels.jl for Bessel function calculations
Replace calls to `SpecialFunctions.besseli(0, ...)` with calls to
`Bessels.besseli0(...)`, adding Bessels.jl
(https://github.com/JuliaMath/Bessels.jl) as a dependency.

When resampling signals with a very low or high arbitrary rate, the call
to `Windows.kaiser` is what dominates.  Profiling this, most of the time
is taken up by the call to `SpecialFunctions.besseli`. This allocates,
but also the computation is just fundamentally slow for these arguments.

Replacing these calls with those to `Bessels.besseli0` gives a factor
of 17 speedup when resampling a 1000-point random trace to a rate of
0.02.

Closes #506.