Fix `setphase!` for pathological cases (#593)
* Fix `setphase!(::Union{FIRInterpolator, FIRRational}, ϕ)` for pathological cases

For odd number of phases `Nφ` and even length of the filter kernel, the old
code would yield `qidx` above `Nφ`, namely equal tp `Nφ+1`, which would result
in a `BoundsError` during `resample`. Rewrite the code to reduce `qidx` to `1`
in this case and increase `inputDeficit` by one instead. All other
constellations should be unaffected.

* Fix `setphase!(::FIRArbitrary, ϕ)` similarly

Here, the problem is less severe as it requires an explcit call to `setphase!`
with a problematic phase which cannot be hit within normal `resample`
operation.